[CHANGE] move Guild, Member Worker to `Guild` Cog
